Genéricos y migración de tipos:   - Crear interfaz CrudService<T, ID> con CRUD completo: listar,     listarPorId, crear, actualizar, eliminar   - ActaService, AutoridadService e InfraccionService extienden     CrudService<T, Long>   - Renombrar métodos (crearActa → crear, obtenerTodasLasActas → listar, etc.)   - Agregar métodos faltantes en cada implementación (listarPorId,     actualizar, eliminar en ActaServiceImpl; listarPorId, actualizar     en AutoridadServiceImpl e InfraccionServiceImpl)   - Migrar PKs de int a Long en Persona, ActaDeConstatacion e Infraccion   - Actualizar repositorios a JpaRepository<T, Long>   - Actualizar @PathVariable a Long   - Corregir DataInitializer: literales DNI con sufijo L   -Limpieza de la migración Swing

+++ src/main/java/com/mycompany/IntegradorMVC/model/ActaDeConstatacion.java
@@ -21,7 +21,7 @@ public class ActaDeConstatacion implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idActa;
+    private Long idActa;
     @OneToMany(cascade = CascadeType.ALL, orphanRemoval = true)
     @JoinColumn(name = "acta_id")
     private List<Infraccion> infracciones = new ArrayList<>();
@@ -70,11 +70,11 @@ public class ActaDeConstatacion implements Serializable {
         this.ruta = ruta;
     }
 
-    public int getIdActa() {
+    public Long getIdActa() {
         return idActa;
     }
 
-    public void setIdActa(int idActa) {
+    public void setIdActa(Long idActa) {
         this.idActa = idActa;
     }
