+++ src/main/java/com/mycompany/IntegradorMVC/repository/AutoridadDeConstatacionRepository.java
@@ -6,9 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AutoridadDeConstatacionRepository extends JpaRepository<AutoridadDeConstatacion, Integer> {
+public interface AutoridadDeConstatacionRepository extends JpaRepository<AutoridadDeConstatacion, Long> {
 
-    boolean existsByDni(int dni);
+    boolean existsByDni(Long dni);
 
-    Optional<AutoridadDeConstatacion> findByDni(int dni);
+    Optional<AutoridadDeConstatacion> findByDni(Long dni);
 }
